added documentation, test in progress
